+++ R/find_random.R
@@ -83,6 +83,10 @@ find_random.afex_aov <- function(x, split_nested = FALSE, flatten = FALSE) {
   if (inherits(x, "brmsfit")) {
     components <- c(components, paste0(.brms_aux_elements(), "_random"))
   }
+  # for glmmTMB, we can have random effects for dispersion component, too
+  if (inherits(x, "glmmTMB")) {
+    components <- c(components, "dispersion_random")
+  }
 
   # check which components we have
   components <- components[vapply(components, function(i) object_has_names(f, i), logical(1))]