+++ R/fish.R
@@ -1,7 +1,20 @@
 #' @docType data
-#' @title Sample data set
+#' @title Sample data set for count models
 #' @name fish
 #' @keywords data
 #'
-#' @description A sample data set, used in tests and some examples.
+#' @description A sample data set, used in tests and some examples. Useful for
+#' testing count models.
+NULL
+
+#' @docType data
+#' @title Sample dataset from the EFC Survey
+#' @name efc_insight
+#' @keywords data
+#'
+#' @description A sample data set, internally used in tests. Consists of 28
+#' variables from 908 observations. The data set is part of the EUROFAMCARE
+#' project, a cross-national survey on informal caregiving in Europe. Useful
+#' when testing on "real-life" data sets, including random missing values. This
+#' data set also has value and variable label attributes.
 NULL