+++ tests/testthat/test-export_table.R
@@ -171,13 +171,12 @@ test_that("export_table, table_width, remove duplicated empty lines", {
   expect_snapshot(print(export_table(out, table_width = 60, empty_line = "-", sep = " | ", remove_duplicates = TRUE)))
   expect_snapshot(print(export_table(out, table_width = 60, empty_line = "-", cross = "+", remove_duplicates = TRUE)))
 
-  skip_if_not_installed("ggeffects")
-  data(efc, package = "ggeffects")
-  out <- datawizard::data_codebook(efc[, 1:4])
+  data(efc_insight, package = "insight")
+  out <- datawizard::data_codebook(efc_insight[, 1:4])
   out$.row_id <- NULL
   expect_snapshot(print(export_table(out, table_width = 60, remove_duplicates = TRUE, empty_line = "-", cross = "+")))
   expect_snapshot(print(export_table(out, table_width = 60, remove_duplicates = FALSE, empty_line = "-", cross = "+")))
-  out <- datawizard::data_codebook(efc[, 1:3])
+  out <- datawizard::data_codebook(efc_insight[, 1:3])
   out$.row_id <- NULL
   expect_snapshot(print(export_table(out, table_width = 60, remove_duplicates = TRUE, empty_line = "-", cross = "+")))
   expect_snapshot(print(export_table(out, table_width = 60, remove_duplicates = FALSE, empty_line = "-", cross = "+")))