+++ tests/testthat/test-glmmTMB.R
@@ -1075,4 +1075,33 @@ test_that("get/find_parameters with dispersion-random", {
     ),
     ignore_attr = TRUE
   )
+
+  # formula and find random works
+  out <- find_formula(m)
+  expect_equal(
+    out,
+    list(
+      conditional = count ~ spp + cover + mined,
+      random = ~ 1 | site,
+      zero_inflated = ~ spp + mined,
+      dispersion = ~DOY,
+      dispersion_random = ~ 1 | site
+    ),
+    ignore_attr = TRUE
+  )
+  out <- find_random(m)
+  expect_identical(out, list(random = "site", dispersion_random = "site"))
+
+  skip_on_cran()
+
+  data(Salamanders, package = "glmmTMB")
+  m <- glmmTMB::glmmTMB(
+    count ~ spp + cover + mined + (1 | site),
+    ziformula = ~ spp + mined,
+    dispformula = ~ DOY + (1 + Wtemp | site),
+    data = Salamanders,
+    family = poisson()
+  )
+  out <- find_random_slopes(m)
+  expect_identical(out, list(dispersion_random = "Wtemp"))
 })